fix(TNLT-8346): Ads not displaying on Android (#492)
Since the Hermes upgrade (#480), the ability to get the source code of a function as a string is broken. That functionality is used when generating the JS to be included in the advert webview.

There is a workaround described in the Github issue: https://github.com/facebook/hermes/issues/114#issuecomment-887106990. But I have had very limited success in getting the `"show source"` directive to work properly. Sometimes it works and sometimes `toString` still returns `function() { [bytecode] }` 🤷 .